+++ src/Extension/Autolink/AutolinkExtension.php
@@ -14,14 +14,13 @@ declare(strict_types=1);
 namespace League\CommonMark\Extension\Autolink;
 
 use League\CommonMark\Environment\ConfigurableEnvironmentInterface;
-use League\CommonMark\Event\DocumentParsedEvent;
 use League\CommonMark\Extension\ExtensionInterface;
 
 final class AutolinkExtension implements ExtensionInterface
 {
     public function register(ConfigurableEnvironmentInterface $environment): void
     {
-        $environment->addEventListener(DocumentParsedEvent::class, new EmailAutolinkProcessor());
-        $environment->addEventListener(DocumentParsedEvent::class, new UrlAutolinkProcessor());
+        $environment->addInlineParser(new EmailAutolinkParser());
+        $environment->addInlineParser(new UrlAutolinkParser());
     }
 }